Move PostRegistrationHelper into ApplicationHelper
There is no PostRegistrationController any more.
